fix: finish keyboard
