+++ src/index.html
@@ -113,15 +113,6 @@
         </div>
         <input list="menu" id="search__field" contenteditable="true" autoComplete="off">
         <div class="countries__table-wrap"></div>
-
-        <!-- <table id="countries__table">
-          <caption id="countries__table-title">Cases by country</caption>
-          <tr class="countries__table-row">
-            <td class="countries__table-cell data"></td>
-            <td class="countries__table-cell name"></td>
-            <td class="countries__table-cell image"><img src="" class="flag-image"></td>
-          </tr>
-        </table> -->
         <div class="button__wrap-down">
           <button class="button__confirmed-countries button button_margin select">CONFIRMED</button>
           <button class="button__recovered-countries button">RECOVERED</button>
