+++ src/keyboardConstants.js
@@ -1,0 +1,596 @@
+const language = {
+  ru: [
+    {
+      small: 'CapsLock',
+      shift: 'CapsLock',
+      code: 'CapsLock',
+    },
+    {
+      small: 'ё',
+      shift: 'Ё',
+      code: 'Backquote',
+    },
+    {
+      small: '1',
+      shift: '!',
+      code: 'Digit1',
+    },
+    {
+      small: '2',
+      shift: '"',
+      code: 'Digit2',
+    },
+    {
+      small: '3',
+      shift: '№',
+      code: 'Digit3',
+    },
+    {
+      small: '4',
+      shift: ';',
+      code: 'Digit4',
+    },
+    {
+      small: '5',
+      shift: '%',
+      code: 'Digit5',
+    },
+    {
+      small: '6',
+      shift: ':',
+      code: 'Digit6',
+    },
+    {
+      small: '7',
+      shift: '?',
+      code: 'Digit7',
+    },
+    {
+      small: '8',
+      shift: '*',
+      code: 'Digit8',
+    },
+    {
+      small: '9',
+      shift: '(',
+      code: 'Digit9',
+    },
+    {
+      small: '0',
+      shift: ')',
+      code: 'Digit0',
+    },
+    {
+      small: '-',
+      shift: '_',
+      code: 'Minus',
+    },
+    {
+      small: '=',
+      shift: '+',
+      code: 'Equal',
+    },
+    {
+      small: 'Backspace',
+      shift: null,
+      code: 'Backspace',
+    },
+    {
+      small: 'й',
+      shift: 'Й',
+      code: 'KeyQ',
+    },
+    {
+      small: 'ц',
+      shift: 'Ц',
+      code: 'KeyW',
+    },
+    {
+      small: 'у',
+      shift: 'У',
+      code: 'KeyE',
+    },
+    {
+      small: 'к',
+      shift: 'К',
+      code: 'KeyR',
+    },
+    {
+      small: 'е',
+      shift: 'Е',
+      code: 'KeyT',
+    },
+    {
+      small: 'н',
+      shift: 'Н',
+      code: 'KeyY',
+    },
+    {
+      small: 'г',
+      shift: 'Г',
+      code: 'KeyU',
+    },
+    {
+      small: 'ш',
+      shift: 'Ш',
+      code: 'KeyI',
+    },
+    {
+      small: 'щ',
+      shift: 'Щ',
+      code: 'KeyO',
+    },
+    {
+      small: 'з',
+      shift: 'З',
+      code: 'KeyP',
+    },
+    {
+      small: 'х',
+      shift: 'Х',
+      code: 'BracketLeft',
+    },
+    {
+      small: 'ъ',
+      shift: 'Ъ',
+      code: 'BracketRight',
+    },
+    {
+      small: 'Enter',
+      shift: null,
+      code: 'Enter',
+    },
+    {
+      small: 'ф',
+      shift: 'Ф',
+      code: 'KeyA',
+    },
+    {
+      small: 'ы',
+      shift: 'Ы',
+      code: 'KeyS',
+    },
+    {
+      small: 'в',
+      shift: 'В',
+      code: 'KeyD',
+    },
+    {
+      small: 'а',
+      shift: 'А',
+      code: 'KeyF',
+    },
+    {
+      small: 'п',
+      shift: 'П',
+      code: 'KeyG',
+    },
+    {
+      small: 'р',
+      shift: 'Р',
+      code: 'KeyH',
+    },
+    {
+      small: 'о',
+      shift: 'О',
+      code: 'KeyJ',
+    },
+    {
+      small: 'л',
+      shift: 'Л',
+      code: 'KeyK',
+    },
+    {
+      small: 'д',
+      shift: 'Д',
+      code: 'KeyL',
+    },
+    {
+      small: 'ж',
+      shift: 'Ж',
+      code: 'Semicolon',
+    },
+    {
+      small: 'э',
+      shift: 'Э',
+      code: 'Quote',
+    },
+    {
+      small: '\\',
+      shift: '/',
+      code: 'Backslash',
+    },
+    {
+      small: 'Shift',
+      shift: null,
+      code: 'ShiftLeft',
+    },
+    {
+      small: '/',
+      shift: '|',
+      code: 'IntlBackslash',
+    },
+    {
+      small: 'я',
+      shift: 'Я',
+      code: 'KeyZ',
+    },
+    {
+      small: 'ч',
+      shift: 'Ч',
+      code: 'KeyX',
+    },
+    {
+      small: 'с',
+      shift: 'С',
+      code: 'KeyC',
+    },
+    {
+      small: 'м',
+      shift: 'М',
+      code: 'KeyV',
+    },
+    {
+      small: 'и',
+      shift: 'И',
+      code: 'KeyB',
+    },
+    {
+      small: 'т',
+      shift: 'Т',
+      code: 'KeyN',
+    },
+    {
+      small: 'ь',
+      shift: 'Ь',
+      code: 'KeyM',
+    },
+    {
+      small: 'б',
+      shift: 'Б',
+      code: 'Comma',
+    },
+    {
+      small: 'ю',
+      shift: 'Ю',
+      code: 'Period',
+    },
+    {
+      small: '.',
+      shift: ',',
+      code: 'Slash',
+    },
+    {
+      small: ' ',
+      shift: null,
+      code: 'Space',
+    },
+    {
+      small: '&larr;',
+      shift: null,
+      code: 'ArrowLeft',
+    },
+    {
+      small: '&uarr;',
+      shift: null,
+      code: 'ArrowUp',
+    },
+    {
+      small: '&darr;',
+      shift: null,
+      code: 'ArrowDown',
+    },
+    {
+      small: '&rarr;',
+      shift: null,
+      code: 'ArrowRight',
+    },
+    {
+      small: 'Ru',
+      shift: 'Ru',
+      code: 'Lang',
+    },
+  ],
+  en: [
+    {
+      small: '`',
+      shift: '~',
+      code: 'Backquote',
+    },
+    {
+      small: '1',
+      shift: '!',
+      code: 'Digit1',
+    },
+    {
+      small: '2',
+      shift: '@',
+      code: 'Digit2',
+    },
+    {
+      small: '3',
+      shift: '#',
+      code: 'Digit3',
+    },
+    {
+      small: '4',
+      shift: '$',
+      code: 'Digit4',
+    },
+    {
+      small: '5',
+      shift: '%',
+      code: 'Digit5',
+    },
+    {
+      small: '6',
+      shift: '^',
+      code: 'Digit6',
+    },
+    {
+      small: '7',
+      shift: '&',
+      code: 'Digit7',
+    },
+    {
+      small: '8',
+      shift: '*',
+      code: 'Digit8',
+    },
+    {
+      small: '9',
+      shift: '(',
+      code: 'Digit9',
+    },
+    {
+      small: '0',
+      shift: ')',
+      code: 'Digit0',
+    },
+    {
+      small: '-',
+      shift: '_',
+      code: 'Minus',
+    },
+    {
+      small: '=',
+      shift: '+',
+      code: 'Equal',
+    },
+    {
+      small: 'Backspace',
+      shift: null,
+      code: 'Backspace',
+    },
+    {
+      small: 'q',
+      shift: 'Q',
+      code: 'KeyQ',
+    },
+    {
+      small: 'w',
+      shift: 'W',
+      code: 'KeyW',
+    },
+    {
+      small: 'e',
+      shift: 'E',
+      code: 'KeyE',
+    },
+    {
+      small: 'r',
+      shift: 'R',
+      code: 'KeyR',
+    },
+    {
+      small: 't',
+      shift: 'T',
+      code: 'KeyT',
+    },
+    {
+      small: 'y',
+      shift: 'Y',
+      code: 'KeyY',
+    },
+    {
+      small: 'u',
+      shift: 'U',
+      code: 'KeyU',
+    },
+    {
+      small: 'i',
+      shift: 'I',
+      code: 'KeyI',
+    },
+    {
+      small: 'o',
+      shift: 'O',
+      code: 'KeyO',
+    },
+    {
+      small: 'p',
+      shift: 'P',
+      code: 'KeyP',
+    },
+    {
+      small: '[',
+      shift: '{',
+      code: 'BracketLeft',
+    },
+    {
+      small: ']',
+      shift: '}',
+      code: 'BracketRight',
+    },
+    {
+      small: 'Enter',
+      shift: null,
+      code: 'Enter',
+    },
+    {
+      small: 'CapsLock',
+      shift: null,
+      code: 'CapsLock',
+    },
+    {
+      small: 'a',
+      shift: 'A',
+      code: 'KeyA',
+    },
+    {
+      small: 's',
+      shift: 'S',
+      code: 'KeyS',
+    },
+    {
+      small: 'd',
+      shift: 'D',
+      code: 'KeyD',
+    },
+    {
+      small: 'f',
+      shift: 'F',
+      code: 'KeyF',
+    },
+    {
+      small: 'g',
+      shift: 'G',
+      code: 'KeyG',
+    },
+    {
+      small: 'h',
+      shift: 'H',
+      code: 'KeyH',
+    },
+    {
+      small: 'j',
+      shift: 'J',
+      code: 'KeyJ',
+    },
+    {
+      small: 'k',
+      shift: 'K',
+      code: 'KeyK',
+    },
+    {
+      small: 'l',
+      shift: 'L',
+      code: 'KeyL',
+    },
+    {
+      small: ';',
+      shift: ':',
+      code: 'Semicolon',
+    },
+    {
+      small: "'",
+      shift: '"',
+      code: 'Quote',
+    },
+    {
+      small: '\\',
+      shift: '|',
+      code: 'Backslash',
+    },
+    {
+      small: 'Shift',
+      shift: null,
+      code: 'ShiftLeft',
+    },
+    {
+      small: '<',
+      shift: '>',
+      code: 'IntlBackslash',
+    },
+    {
+      small: 'z',
+      shift: 'Z',
+      code: 'KeyZ',
+    },
+    {
+      small: 'x',
+      shift: 'X',
+      code: 'KeyX',
+    },
+    {
+      small: 'c',
+      shift: 'C',
+      code: 'KeyC',
+    },
+    {
+      small: 'v',
+      shift: 'V',
+      code: 'KeyV',
+    },
+    {
+      small: 'b',
+      shift: 'B',
+      code: 'KeyB',
+    },
+    {
+      small: 'n',
+      shift: 'N',
+      code: 'KeyN',
+    },
+    {
+      small: 'm',
+      shift: 'M',
+      code: 'KeyM',
+    },
+    {
+      small: ',',
+      shift: '<',
+      code: 'Comma',
+    },
+    {
+      small: '.',
+      shift: '>',
+      code: 'Period',
+    },
+    {
+      small: '/',
+      shift: '?',
+      code: 'Slash',
+    },
+    {
+      small: ' ',
+      shift: null,
+      code: 'Space',
+    },
+    {
+      small: '&larr;',
+      shift: null,
+      code: 'ArrowLeft',
+    },
+    {
+      small: '&uarr;',
+      shift: null,
+      code: 'ArrowUp',
+    },
+    {
+      small: '&darr;',
+      shift: null,
+      code: 'ArrowDown',
+    },
+    {
+      small: '&rarr;',
+      shift: null,
+      code: 'ArrowRight',
+    },
+    {
+      small: 'En',
+      shift: 'En',
+      code: 'Lang',
+    },
+  ],
+};
+
+const rowsOrder = [
+  ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal'],
+  ['KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Backspace'],
+  ['CapsLock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Backslash', 'Enter'],
+  ['ShiftLeft', 'IntlBackslash', 'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ArrowUp'],
+  ['Lang', 'Space', 'ArrowLeft', 'ArrowDown', 'ArrowRight'],
+];
+
+  export { language, rowsOrder };
