+++ src/styles.css
@@ -572,7 +572,7 @@ background-color: rgb(39, 0, 0);
     max-height: 535px;
     position: absolute;
     background-color: white;
-    top: 30%;
+    top: 25%;
     color: black;
     width: inherit;
     overflow: auto;
@@ -596,11 +596,13 @@ background-color: rgb(39, 0, 0);
 }
 
 .dark {
-    background-color: rgba(14, 13, 13);
+	background-color: rgba(14, 13, 13);
+	z-index: 999;
 }
 
 .black {
-    background-color: rgb(77, 77, 77);
+	background-color: rgb(77, 77, 77);
+	z-index: 999;
 }
 
 .flag-image {
@@ -608,19 +610,59 @@ background-color: rgb(39, 0, 0);
 }
 
 .keyboard {
-    display: grid;
+    display: none;
     color: white;
     position: relative;
     overflow: hidden;
     transition: 1s linear;
     font-size: 18px;
     left: 0;
-    bottom: 0;
+    bottom: 91vh;
+    z-index: 999;
+    background-color: #080808;
+    width: 74vw;
+    height: 40vh;
+	box-shadow: 0 0 20px 13px;
+	text-align: center;
 }
 
 .keyboard__row {
     grid-template-columns: repeat(13, 1fr);
-    display: grid;
+	display: grid;
+	height: 8vh;
+}
+
+.keyboard__key {
+	border: 1px solid white;
+}
+.keyboard__key:hover {
+	background-color:  white;
+	cursor: pointer;
+}
+
+.keyboard__key:active {
+	box-shadow: inset 0 0 5px 5px grey;
+	cursor: pointer;
+}
+
+.off {
+	display: none;
+}
+
+.on {
+	display: grid;
+}
+
+.active {
+	background-color: rgb(160, 4, 4);
+}
+
+.sub {
+	color: rgb(160, 4, 4);
+}
+
+.letter {
+	color: rgb(77, 77, 77);
 }
 
 /* footer */
